+++ starLib/src/main/java/edu/illinois/mitra/starl/motion/MiniDroneBTI.java
@@ -133,7 +133,7 @@ public class MiniDroneBTI implements DroneBTI, ARDiscoveryServicesDevicesListUpd
         }
     }
 
-    public void setControlInput(float yaw, float pitch, float roll, float gaz) {
+    public void setControlInput(double yaw, double pitch, double roll, double gaz) {
         setYaw(yaw);
         setPitch(pitch);
         setRoll(roll);
