+++ starLib/src/main/java/edu/illinois/mitra/starl/motion/o3DRController.java
@@ -420,7 +420,7 @@ public class o3DRController extends AppCompatActivity implements DroneBTI, Tower
         });
     }
 
-    public void setControlInput(float yaw, float pitch, float roll, float gaz) {
+    public void setControlInput(double yaw, double pitch, double roll, double gaz) {
         setYaw(yaw);
         setPitch(pitch);
         setRoll(roll);
